:construction: implement news local source via room /sorry for big commit

+++ data/room/src/main/java/com/yoloroy/room/news/model/ArticleWithId.kt
@@ -2,7 +2,7 @@ package com.yoloroy.room.news.model
 
 import com.yoloroy.retrofit.news.model.Article
 
-class ArticleWithId(val id: Int, private val article: Article) : Article by article {
+class ArticleWithId(val id: Long, private val article: Article) : Article by article {
 
     override fun short() = NewsShortWithId(id, super.short())
     override fun details() = NewsDetailsWithId(id, super.details())
