+++ data/room/src/main/java/com/yoloroy/room/news/model/NewsDetailsWithId.kt
@@ -2,4 +2,4 @@ package com.yoloroy.room.news.model
 
 import com.yoloroy.domain.model.NewsDetails
 
-class NewsDetailsWithId(val id: Int, private val newsDetails: NewsDetails) : NewsDetails by newsDetails
+class NewsDetailsWithId(val id: Long, private val newsDetails: NewsDetails) : NewsDetails by newsDetails
