+++ data/room/src/main/java/com/yoloroy/room/news/model/NewsShortWithId.kt
@@ -2,4 +2,4 @@ package com.yoloroy.room.news.model
 
 import com.yoloroy.domain.model.NewsShort
 
-class NewsShortWithId(val id: Int, private val newsShort: NewsShort) : NewsShort by newsShort
+class NewsShortWithId(val id: Long, private val newsShort: NewsShort) : NewsShort by newsShort
